+++ packages/sdk/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+- fix `SparseMerkleProver` bug where empty tree has incorrect root
 - fix minor bug where SDK would override `executionGasLimit` or `maxNumRefunds` even if it was set in opRequest
 - make commitment tree quaternary:
   - update `proof/joinsplit` and `proof/subtreeupdate` with new circuit input shapes